% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholesky.R
\name{chol}
\alias{chol}
\title{A complete Cholesky factorization}
\usage{
chol(A)
}
\arguments{
\item{A}{A symmetric positive definite matrix to be factored. No default value.}
}
\description{
The function computes Cholesky factor as a lower triangular matrix.
}
\examples{
A = matrix(c(1,-1,-1,1), ncol=2)
L = chol(A)
}
\keyword{Cholesky}
