% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.linalg.R
\name{cond}
\alias{cond}
\title{Condition number}
\usage{
cond(A)
}
\arguments{
\item{A}{A non-singular matrix. No default value.}
}
\description{
Computes the condition number of the given matrix.
}
\examples{
A = diag(c(1,2,3))
k = cond(A)
}
\keyword{condition}
\keyword{number}
