% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.linalg.R
\name{det.tridiag}
\alias{det.tridiag}
\title{Determinant of a tridiagonal matrix}
\usage{
det.tridiag(H)
}
\arguments{
\item{H}{A tridiagonal matrix whose determinant is to be computed. No default value.}
}
\description{
Computes the determinant of a tridiagonal matrix using continuants three term relation.
}
\examples{
H = tridiag(c(1,2,3), c(4,5,6), c(7,8,9))
det.tridiag(H)
}
\keyword{determinant}
