% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholesky.R
\name{ichol}
\alias{ichol}
\title{An Incomplete Cholesky factorization}
\usage{
ichol(A, diag.comp = "aj", alpha = 1)
}
\arguments{
\item{A}{A symmetric positive definite matrix to be factored. No default value.}

\item{diag.comp}{A type of diagonal compensation. Either "aj" (default) for Ajiz-Jennings,
"as" for absolute shift (diag(A) = diag(A) + alpha), "rs" for relative shift
(diag(A) = diag(A) + alpha * diag(A)).}

\item{alpha}{Value for diagonal compensation. Default is 1. For matrices not diagonally dominant,
small values of the parameter could lead to negative pivots. When such are encountered,
the pivot is replaced by 1.}
}
\description{
The function compues IC(0) factorization as a lower triangular matrix.
}
\examples{
A = matrix(c(1,-1,-1,1), ncol=2)
L = ichol(A, diag.comp="as", alpha=1)
}
\keyword{Cholesky,}
\keyword{Incomplete}
\keyword{LU}
