% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos}
\alias{lanczos}
\title{Lanczos algorithm}
\usage{
lanczos(A, b, x = rnorm(length(b)), m = length(b), tol = 1e-07)
}
\arguments{
\item{A}{A symmetric positive definite matrix for the system Ax = b. No default value.}

\item{b}{A right hand side vector in Ax = b. No default value.}

\item{x}{An initial guess which produces the first residual r=b-Ax. Default is a random vector.}

\item{m}{Maximum number of iterations to be run. Default is length(b).}

\item{tol}{The relative tolerance for residual norms. Default is 1e-7.}
}
\description{
The function solve a symmetric positive definite system by factoring the matrix into A=VHV^T form.
Here, V is matrix containing Krylov orthogonal vectors, and H is a tridiagonal matrix with Lanczos
coefficients.
}
\examples{
A = matrix(rnorm(100^2), 100)
A = t(A)\%*\%A
b = rnorm(100)
res = lanczos(A, b, 80)
plot(solve(A, b), t="l")
lines(res$x, col="red")
}
\keyword{Krylov,}
\keyword{Lanczos,}
\keyword{tridiagonalization}
