% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.precond.R
\name{planczos}
\alias{planczos}
\title{Preconditioned Lanczos algorithm}
\usage{
planczos(A, b, m = length(b), tol = 1e-07, diag.comp = "aj", alpha = 1, L)
}
\arguments{
\item{A}{A symmetric positive definite matrix for the system Ax = b. No default value.}

\item{b}{A right hand side vector in Ax = b. No default value.}

\item{m}{Maximum number of iterations to be run. Default is length(b).}

\item{tol}{The relative tolerance for residual norms. Default is 1e-7.}

\item{diag.comp}{A type of diagonal compensation for IC(0) preconditioning. Default is "aj".}

\item{alpha}{The scale of diagonal compensation. Default is 1.}

\item{L}{The preconditioner to supply. If missing, the IC(0) is above parameters is computed.}
}
\description{
The function solve a symmetric positive definite system by split preconditioning the system Ax = b.
Then it is handled by a plain Lanczos algorithm.
}
\examples{
A = matrix(rnorm(100^2), 100)
A = t(A)\%*\%A
b = rnorm(100)
res = planczos(A, b)
plot(solve(A, b), t="l")
lines(res$x, col="red")
}
\keyword{Krylov,}
\keyword{Lanczos,}
\keyword{tridiagonalization}
